Separate out transition generation from ProducerSourceBase
Moved the generation of new transition id into a new class IDGeneratorSourceBase. This allows EmptySource to reuse that code but not pay the time penalties for the overhead involved in being a Producer.
