+++ FWCore/Modules/src/EmptySource.cc
@@ -2,10 +2,11 @@
 #include "FWCore/Framework/interface/InputSourceMacros.h"
 #include "FWCore/ParameterSet/interface/ConfigurationDescriptions.h"
 #include "FWCore/ParameterSet/interface/ParameterSetDescription.h"
-#include "FWCore/Sources/interface/ProducerSourceBase.h"
+#include "FWCore/Framework/interface/InputSource.h"
+#include "FWCore/Sources/interface/IDGeneratorSourceBase.h"
 
 namespace edm {
-  class EmptySource : public ProducerSourceBase {
+  class EmptySource : public IDGeneratorSourceBase<InputSource> {
   public:
     explicit EmptySource(ParameterSet const&, InputSourceDescription const&);
     ~EmptySource() override;
@@ -13,22 +14,24 @@ namespace edm {
 
   private:
     bool setRunAndEventInfo(EventID& id, TimeValue_t& time, edm::EventAuxiliary::ExperimentType&) override;
-    void produce(Event&) override;
+    void readEvent_(edm::EventPrincipal&) override;
   };
 
   EmptySource::EmptySource(ParameterSet const& pset, InputSourceDescription const& desc)
-      : ProducerSourceBase(pset, desc, false) {}
+      : IDGeneratorSourceBase<InputSource>(pset, desc, false) {}
 
   EmptySource::~EmptySource() {}
 
   bool EmptySource::setRunAndEventInfo(EventID&, TimeValue_t&, edm::EventAuxiliary::ExperimentType&) { return true; }
 
-  void EmptySource::produce(edm::Event&) {}
+  void EmptySource::readEvent_(edm::EventPrincipal& e) {
+    doReadEvent(e, [](auto const&) {});
+  }
 
   void EmptySource::fillDescriptions(ConfigurationDescriptions& descriptions) {
     ParameterSetDescription desc;
     desc.setComment("Creates runs, lumis and events containing no products.");
-    ProducerSourceBase::fillDescription(desc);
+    IDGeneratorSourceBase<InputSource>::fillDescription(desc);
     descriptions.add("source", desc);
   }
 }  // namespace edm
